Fix API route: move OpenAI handler to /api/generate-metadata
