+++ pages/tools/generate-metadata.js
@@ -1,29 +1,9 @@
-import { OpenAI } from "openai";
+import MetadataGenerator from "@/components/MetadataGenerator";
 
-const openai = new OpenAI({
-  apiKey: process.env.OPENAI_API_KEY,
-});
-
-export default async function handler(req, res) {
-  if (req.method !== "POST") return res.status(405).end("Method Not Allowed");
-
-  const { prompt } = req.body;
-
-  try {
-    const chat = await openai.chat.completions.create({
-      model: "gpt-4",
-      messages: [
-        { role: "system", content: "You are a helpful assistant for game content generation." },
-        { role: "user", content: prompt },
-      ],
-      temperature: 0.7,
-      max_tokens: 600,
-    });
-
-    const output = chat.choices[0].message.content;
-    res.status(200).json({ output });
-  } catch (err) {
-    console.error("OpenAI error:", err);
-    res.status(500).json({ error: "API call failed" });
-  }
+export default function AdminToolsPage() {
+  return (
+    <div className="p-6">
+      <MetadataGenerator />
+    </div>
+  );
 }
